requirements.txt
python-docx==1.1.2
docx2pdf==0.1.8
PyPDF2==3.0.1
customtkinter==5.2.1
Pillow>=10.0.0
secure-smtplib==0.1.1
email-validator==2.1.0
python-dotenv==1.0.0
